% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase-package.R
\docType{package}
\name{freebase}
\alias{freebase}
\alias{freebase-package}
\title{A 'usethis'-like Package for Base Pseudo-equivalents of 'tidyverse' Code}
\description{
The 'tidyverse' is awesome, but can take a bit compile on systems
where there are no pre-built binary packages. Methods are provided which use
the facilities of the 'usethis' package to snap-in base versions of useful
'tidyverse' functions that are mostly equivalent (some are more complete
than others). The base R counterpart functions will likely be slower than
the 'tidyverse' equivalents but using them will decrease 'Imports'
dependencies.\cr\crUse deliberately and with caution.
}
\author{
Bob Rudis (bob@rud.is)
}
