% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-detect.R
\name{use_detect}
\alias{use_detect}
\title{Use base R versions of \code{purrr} \code{detect} functions}
\usage{
use_detect(save_as = "R/utils-detect.R", open = TRUE)
}
\arguments{
\item{save_as}{File path and name. Defaults to "\code{R/utils-detect.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates base R versions of \code{purrr} \code{detect} (et al).
}
}
\details{
These support formula functions (i.e. \code{~{}}).
}
