% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-keepers.R
\name{use_keepers}
\alias{use_keepers}
\title{Use base R versions of \code{purrr} \code{keep} functions}
\usage{
use_keepers(save_as = "R/utils-keepers.R", open = TRUE)
}
\arguments{
\item{save_as}{File path and name. Defaults to "\code{R/utils-keepers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates base R versions of \code{purrr} \code{keep} fuctions.
}
}
\details{
These support formula functions (i.e. \code{~{}}).
}
