% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-tidylite.R
\name{use_tidylite}
\alias{use_tidylite}
\title{Use base R versions of \code{tidyr} functions}
\usage{
use_tidylite(save_as = "R/utils-tidylite.R", open = TRUE)
}
\arguments{
\item{save_as}{File path and name. Defaults to "\code{R/utils-tidylite.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates base R versions of \code{tidyr} functions
}
}
