% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-walkers.R
\name{use_walkers}
\alias{use_walkers}
\title{Use base R versions of \code{purrr} \code{walk} functions}
\usage{
use_walkers(save_as = "R/utils-walkers.R", open = TRUE)
}
\arguments{
\item{save_as}{File path and name. Defaults to "\code{R/utils-walkers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates base R versions of \code{purrr} \code{walk} functions
}
}
\details{
These support formula functions (i.e. \code{~{}}).
}
