% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-infix-helpers.R
\name{use_infix_helpers}
\alias{use_infix_helpers}
\title{Use infix-helpers}
\usage{
use_infix_helpers(save_as = "R/utils-infix-helpers.R", open = TRUE)
}
\arguments{
\item{save_as}{Where to save/what to name the file. Defaults to "\code{R/utils-infix-helpers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates \code{R/utils-infix-helpers.R} with base R \code{\%XXX\%} helpers
}
}
