% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-isers.R
\name{use_isers}
\alias{use_isers}
\title{Use "is_" functions}
\usage{
use_isers(save_as = "R/utils-isers.R", open = TRUE)
}
\arguments{
\item{save_as}{Where to save/what to name the file. Defaults to "\code{R/utils-isers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates \code{R/utils-isers.R} filled with \code{is_...} testors
}
}
