% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-mappers.R
\name{use_mappers}
\alias{use_mappers}
\title{Use base R versions of \code{purrr} \code{map} functions}
\usage{
use_mappers(save_as = "R/utils-mappers.R", open = TRUE)
}
\arguments{
\item{save_as}{File path and name. Defaults to "\code{R/utils-mappers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates base R versions of \code{purrr} \code{map} functions.
}
}
\details{
These support formula functions (i.e. \code{~{}}).
}
