% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-safely.R
\name{use_safely}
\alias{use_safely}
\title{Use base-ified equivalents of "safely"-ifiers}
\usage{
use_safely(save_as = "R/utils-safely.R", open = TRUE)
}
\arguments{
\item{save_as}{Where to save/what to name the file. Defaults to "\code{R/utils-safely.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates \code{R/utils-safely.R} with base R pseudo-equivalents for \code{purrr} \code{safely} (et al).
}
}
