% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-tidylite.R
\name{use_tidylite}
\alias{use_tidylite}
\title{Use base R versions of \code{tidyr} \code{gather}/\code{spread}/\code{unite}}
\usage{
use_tidylite(save_as = "R/utils-tidylite.R", open = TRUE)
}
\arguments{
\item{save_as}{Where to save/what to name the file. Defaults to "\code{R/utils-tidylite.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates \code{R/utils-tidylite.R} containing \code{gatherlite()}, \code{spreadlite()} and \code{unitelite()}
}
}
