% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-walkers.R
\name{use_walkers}
\alias{use_walkers}
\title{Use base-ified equivalents of keep/discard/compact}
\usage{
use_walkers(save_as = "R/utils-walkers.R", open = TRUE)
}
\arguments{
\item{save_as}{Where to save/what to name the file. Defaults to "\code{R/utils-walkers.R}"}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
\itemize{
\item Creates \code{R/utils-walkers.R} with base R pseudo-equivalents for \code{purrr}
\code{walk}/\code{walk2}
}
}
\details{
These support formula functions (i.e. \code{~{}}).
}
